chore: Export color conversion logic
- This can help us derive color roles from an image (ie: by estimating the direction we want to take the colors based on its brightness).
